modbus_(rtu|tcp): remove incorrect 'requires' key from setup args. Fix #10 and fix #22.
The 'python-' prefix of the 'pymodbus' entry was not correct: only the debian
package has that prefix. It caused the setup phase to fail, which made the
robotiq docjob error out. It was also the reason for 'catkin_make install' not
completing.

instead of correcting the entry, the entire 'requires' key was removed from
the setup args. Catkin packages should not be installed using the 'setup.py'
file, and the dependency on 'python-modbus' is already expressed in the pkg
manifest.