Possibility to choose  to spawn finger_pads
Since not everyone is using standard Finger Pads  that got added on the 3rd of December.
I added the option to not spawn them, while spawning the gripper via.  <xacro:robotiq_arg2f_140 prefix="${prefix}gripper_"  />
To do so just add the parameter pad="0". eg. <xacro:robotiq_arg2f_140 prefix="${prefix}gripper_" pad="0" />
To keep it backward compatible, the parameter is on default 1. So if no parameter is set it just spawns them as before.